requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6